# Some performance optimizations to the MetricHolder