--- a/metric/held_counter.go
+++ b/metric/held_counter.go
@@ -1,0 +1,43 @@
+package metric
+
+import (
+	"time"
+
+	"github.com/prometheus/client_golang/prometheus"
+)
+
+type HeldCounter struct {
+	*heldMetric[prometheus.Counter]
+}
+
+func (h HeldCounter) Inc() {
+	h.metric.Inc()
+	h.updateUsage()
+}
+
+func (h HeldCounter) Add(v float64) {
+	h.metric.Add(v)
+	h.updateUsage()
+}
+
+type HeldCounterVec struct {
+	store *heldMetricsStore[prometheus.Counter]
+	vec   *prometheus.CounterVec
+}
+
+func NewHeldCounterVec(cv *prometheus.CounterVec) HeldCounterVec {
+	return HeldCounterVec{
+		vec:   cv,
+		store: newHeldLabelsStore[prometheus.Counter](),
+	}
+}
+
+func (h HeldCounterVec) WithLabelValues(lvs ...string) HeldCounter {
+	return HeldCounter{
+		heldMetric: h.store.GetOrCreate(lvs, h.vec.WithLabelValues),
+	}
+}
+
+func (h HeldCounterVec) ReleaseOldMetrics(holdDuration time.Duration) {
+	h.store.ReleaseOldMetrics(holdDuration, h.vec)
+}