--- a/metric/held_histogram.go
+++ b/metric/held_histogram.go
@@ -1,0 +1,40 @@
+package metric
+
+import (
+	"time"
+
+	"github.com/prometheus/client_golang/prometheus"
+)
+
+type HeldHistogram struct {
+	*heldMetric[prometheus.Histogram]
+}
+
+func (h HeldHistogram) Observe(v float64) {
+	h.metric.Observe(v)
+	h.updateUsage()
+}
+
+type HeldHistogramVec struct {
+	store *heldMetricsStore[prometheus.Histogram]
+	vec   *prometheus.HistogramVec
+}
+
+func NewHeldHistogramVec(hv *prometheus.HistogramVec) HeldHistogramVec {
+	return HeldHistogramVec{
+		vec:   hv,
+		store: newHeldLabelsStore[prometheus.Histogram](),
+	}
+}
+
+func (h HeldHistogramVec) WithLabelValues(lvs ...string) HeldHistogram {
+	return HeldHistogram{
+		heldMetric: h.store.GetOrCreate(lvs, func(s ...string) prometheus.Histogram {
+			return h.vec.WithLabelValues(s...).(prometheus.Histogram)
+		}),
+	}
+}
+
+func (h HeldHistogramVec) ReleaseOldMetrics(holdDuration time.Duration) {
+	h.store.ReleaseOldMetrics(holdDuration, h.vec)
+}