--- a/metric/holder.go
+++ b/metric/holder.go
@@ -7,7 +7,7 @@ import (
 )
 
 type heldMetricVec interface {
-	update(holdDuration time.Duration)
+	ReleaseOldMetrics(holdDuration time.Duration)
 }
 
 type Holder struct {
@@ -15,7 +15,11 @@ type Holder struct {
 	heldMetrics  []heldMetricVec
 }
 
+// NewHolder returns new metric holder. The holdDuration must be more than 1m.
 func NewHolder(holdDuration time.Duration) *Holder {
+	if holdDuration < time.Minute {
+		panic("hold duration must be greater than 1m")
+	}
 	return &Holder{
 		holdDuration: holdDuration,
 		heldMetrics:  make([]heldMetricVec, 0),
@@ -23,30 +27,30 @@ func NewHolder(holdDuration time.Duration) *Holder {
 }
 
 func (h *Holder) Maintenance() {
-	h.updateMetrics()
+	h.releaseOldMetrics()
 }
 
-func (h *Holder) AddCounterVec(counterVec *prometheus.CounterVec) *HeldCounterVec {
-	hcv := newHeldCounterVec(counterVec)
+func (h *Holder) AddCounterVec(counterVec *prometheus.CounterVec) HeldCounterVec {
+	hcv := NewHeldCounterVec(counterVec)
 	h.heldMetrics = append(h.heldMetrics, hcv)
 	return hcv
 }
 
-func (h *Holder) AddGaugeVec(gaugeVec *prometheus.GaugeVec) *HeldGaugeVec {
-	hgv := newHeldGaugeVec(gaugeVec)
+func (h *Holder) AddGaugeVec(gaugeVec *prometheus.GaugeVec) HeldGaugeVec {
+	hgv := NewHeldGaugeVec(gaugeVec)
 	h.heldMetrics = append(h.heldMetrics, hgv)
 	return hgv
 }
 
-func (h *Holder) AddHistogramVec(histogramVec *prometheus.HistogramVec) *HeldHistogramVec {
-	hhv := newHeldHistogramVec(histogramVec)
+func (h *Holder) AddHistogramVec(histogramVec *prometheus.HistogramVec) HeldHistogramVec {
+	hhv := NewHeldHistogramVec(histogramVec)
 	h.heldMetrics = append(h.heldMetrics, hhv)
 	return hhv
 }
 
-// updateMetrics delete old metrics, that aren't in use since last update.
-func (h *Holder) updateMetrics() {
+// releaseOldMetrics delete old metric labels, that aren't in use since last update.
+func (h *Holder) releaseOldMetrics() {
 	for i := range h.heldMetrics {
-		h.heldMetrics[i].update(h.holdDuration)
+		h.heldMetrics[i].ReleaseOldMetrics(h.holdDuration)
 	}
 }