--- a/xtime/xtime.go
+++ b/xtime/xtime.go
@@ -6,11 +6,11 @@ import (
 )
 
 func init() {
-	setNowTime(time.Now())
+	SetNowTime(time.Now().UnixNano())
 	ticker := time.NewTicker(updateTimeInterval)
 	go func() {
 		for t := range ticker.C {
-			setNowTime(t)
+			SetNowTime(t.UnixNano())
 		}
 	}()
 }
@@ -27,6 +27,12 @@ func GetInaccurateTime() time.Time {
 	return time.Unix(0, GetInaccurateUnixNano())
 }
 
-func setNowTime(t time.Time) {
-	nowTime.Store(t.UnixNano())
+// SetNowTime tests the current time.
+// Function should be used only in tests.
+//
+// An alternative to this approach is to store and redefine
+// a function through the fields of the tested struct.
+// But in this case, the inlining function GetInaccurateUnixNano is lost.
+func SetNowTime(unixNano int64) {
+	nowTime.Store(unixNano)
 }