metricholder takes up a significant amount of CPU on our profile. This PR adds several optimizations. The main optimizations are:

1) Calculating prometheus.MetricVec.WithLabelValues() once, not for every .Inc/.Set/etc
2) Optimization of hash calculation when there is only 1 label
3) Using xxhash for hashing
4) Reduce the amount of memory allocation in the heap
5) Update lastUsage atomic every 10 seconds, not every .Inc/.Set/etc

(sorted by decreasing impact on the benchmark)

old:
```
$ go test -bench='Benchmark(MetricHolder)' -run=XXX -benchmem ./...
goos: darwin
goarch: arm64
pkg: github.com/ozontech/file.d/metric
BenchmarkMetricHolder/l1-8               3806335               299.4 ns/op           240 B/op          6 allocs/op
BenchmarkMetricHolder/l1_l2-8            2969712               404.7 ns/op           288 B/op          6 allocs/op
BenchmarkMetricHolder/l1_l2_l3-8         2462668               487.5 ns/op           336 B/op          6 allocs/op
PASS
ok      github.com/ozontech/file.d/metric       5.006s
```

new:
```
$ go test -bench='Benchmark(MetricHolder)' -run=XXX -benchmem ./...
goos: darwin
goarch: arm64
pkg: github.com/ozontech/file.d/metric
BenchmarkMetricHolder/l1-8              19625188                60.72 ns/op            0 B/op          0 allocs/op
BenchmarkMetricHolder/l1_l2-8           16686858                71.55 ns/op            0 B/op          0 allocs/op
BenchmarkMetricHolder/l1_l2_l3-8        12369921                97.69 ns/op            0 B/op          0 allocs/op
PASS
ok      github.com/ozontech/file.d/metric       5.217s
```


prom:
```
$ go test -bench='Benchmark(Prom)' -run=XXX -benchmem ./...
goos: darwin
goarch: arm64
pkg: github.com/ozontech/file.d/metric
BenchmarkPromVec/l1-8            7012634               165.3 ns/op            48 B/op          3 allocs/op
BenchmarkPromVec/l1_l2-8         4787998               250.3 ns/op            96 B/op          3 allocs/op
BenchmarkPromVec/l1_l2_l3-8      3826999               316.3 ns/op           144 B/op          3 allocs/op
PASS
ok      github.com/ozontech/file.d/metric       4.440s
```

TODO: 

- [X] Tests
- [X] Benchmark prometheus.MetricVec
- [X] Benchmark result old/new